Refactor: extract PhotoTaker service from UrgeSaver
